Remove low level class resin.Token since all features are now merged into resin.Auth
Change-Type: minor
