+++ resin/__init__.py
@@ -8,7 +8,6 @@ hesitate to open an issue in GitHub, we'll be happy to help.
 
 from .base_request import BaseRequest
 from .auth import Auth
-from .token import Token
 from .logs import Logs
 from .settings import Settings
 from .models import Models
